virtio: correctly prepend net_hdr before transmitting packet
Instead of assuming there is room for a header in front of the packet, use
the packet prepending constructor to check for it and allocate it if missing.

This requires splitting the waiting into two parts: first acquiring the
packet, and only then, when we know how many fragments it has, waiting
for virtio dscriptors.  It also requires that we prepare a full buffer_chain
rather than one with exactly one fragment.