core: special-case deleter for raw memory
In many cases, a deleter is used to protect raw memory (e.g. a char array,
not something with a destructor).  In that case we can simply free() it,
so, the deleter need not remember which destructor needs to be called.

It does need to remember whether it's a raw object or not, so we take over
the least significant bit and use it as a marker, and store the pointer
to the object in the deleter, instead of using a proxy impl object to
control actual deletion.

If the deleter is subsequently share()d, we have to convert it back to
the standard form, since the reference count lives in the impl object.