smp: asynchronous functions
Allow functions run on another cpu to return a future.  In that case, the
result (or exception) is only returned when the future is resolved.

Note this has the potential for livelocks:

A: smp::submit_to(B, [] {
       ...
       return smp::submit_to(A, [] {
           ...
       });
    });

If this and its mirror image (B sending to A) happen concurrently, and if
both the A->B and B->A queue become full, the system will not be able to
make forward progress.  This can be fixed by making the inner submit_to()
use a different queue.