xen: only add features to feature array
We are adding everything we read into the features array. Because in the
destructor we will remove everything in the features list, we'll end up
removing more than we should. Things like the mac address, handle, etc, should
never be deleted.

This is not a problem for OSv because usually, after the destructor is called,
the whole guest is down. But for userspace, the network card is left there,
but will cease to work if we delete too much.

After we do that with the _features array - it's original intent, it becomes
reduntant with features nack.

Signed-off-by: Glauber Costa <glommer@cloudius-systems.com>