dpdk: rx checksum offloading
If the card supports this (and usually, it does), enable rx checksum
offloading by the card, and avoid calculating the checksums ourselves.

With rx checksum offloading, the card checks in incoming packets the
IP header checksum and the L4 (TCP or UDP) checksum, and gives us a
flag when one of them is wrong, meaning that we do not need to do these
calculations ourselves.

This patch improves memcached performance on my setup by almost 3%.

Signed-off-by: Nadav Har'El <nyh@cloudius-systems.com>